Improve error messages for invalid comparisons
The expression `'bar' < 'foo'` used to display an error that said
"expected str, found str", which is certainly quite confusing.  This
commit ensures that:

1. When an operation is valid (i.e., comparison between ints or
floats, equality between strings), but the operands' types don't match
an error message to that effect (expected X, found Y) is displayed;
2. When an operation is invalid (e.g., using the less-than operator on
strings or any comparison operator on lists), the error message that
the operation is invalid.