Fix longest-match rule bug in scanner
The longest-match rule was not respected by the scanner: given the
input "one_direction", the scanner would return two tokens, the
keyword "one" and the variable "_direction".

This commit fixes the issue by removing the usage of .looking_at() for
keywords (which short-circuited the longest-match).  Instead, when a
string looks like a variable name, we consume it entirely and after we
have all the characters we check if it's a keyword or a variable.